Create Pymgur.py
A script which, when run, will take the image on the clipboard, upload it to Imgur, then copy a direct link to the image to the clipboard automatically. This makes sharing images over any messaging service easy. Use the Pythonista Homescreen Shortcut Maker ( http://omz-software.com/pythonista/shortcut/ ) to make this a one click process.